Updating SendReceiveScheduledMessage timeline.
Previously, the message was scheduled to be delivered 15 seconds in the future. Service Bus has a schedule granularity more along the lines of a 1-minute window. For that reason, I update this test to take 3 minutes. From there, I had to update some of the other timeout logic in the test harness to allow a test that took more than 1 minute.