Use link caching for $management operations (renew locks) (#248)
Make it so we use a single management link (most of the time) when doing dispositions or lock renewal (or other management related operations).

Also, added in a simple stress test. It's a bit manual at the moment, but it initiates 1000 concurrent renew lock calls at once, so it's a decent test of our response routing.